UPDATED GRAPHICS
Updated the play, ranking and rate buttons
Updated the dashboard
Updated Miley's hammer
Updated the beams
Updated the wrecking ball

I added a platform.png and a background.png to the folder. They'll have
to be added to the game.

I changed the sizes of a lot of the files by small amounts